Перенос логики в контроллер
