+++ src/components/no-tasks.js
@@ -9,7 +9,7 @@ const createNoTasksTemplate = () => {
   `;
 };
 
-export default class NoTasks extends AbstractComponent {
+export default class NoTasksComponent extends AbstractComponent {
   getTemplate() {
     return createNoTasksTemplate();
   }
